Split up main.go and rename core to p2p
